Print extension names instead of count (#293)
More helpful if no details are provided.
